study01/web-application-server/src/test/java/http/HttpRequestTest.java
package http;

import org.junit.Test;
import http.HttpRequest;
import http.HttpResponse;

import java.io.*;

import static org.junit.Assert.assertEquals;

public class HttpRequestTest {
    private String testDirectory = "./src/test/resources/";

    @Test
    public void request_GET() throws Exception {
        InputStream in = new FileInputStream(new File(testDirectory + "Http_GET"));
        HttpRequest request = new HttpRequest(in);

        assertEquals("GET", request.getMethod());
        assertEquals("/user/create", request.getPath());
        assertEquals("keep-alive", request.getHeader("Connection"));
        assertEquals("dvno", request.getParameter("userId"));
    }
}
